+++ mobgap/utils/array_handling.py
@@ -85,14 +85,17 @@ class MultiGroupBy:
 
     _primary_groupby: pd.core.groupby.DataFrameGroupBy
     _secondary_groupbys: list[pd.core.groupby.DataFrameGroupBy]
+    _kwargs: dict[str, Any]
 
     def __init__(
         self,
         primary_df: pd.DataFrame,
         secondary_dfs: Union[pd.DataFrame, list[pd.DataFrame]],
         groupby: Union[str, list[str]],
+        **kwargs: Unpack[dict[str, Any]],
     ) -> None:
         groupby_as_list = [groupby] if isinstance(groupby, str) else groupby
+        self._kwargs = kwargs
 
         primary_index_cols = primary_df.index.names
         if not set(groupby_as_list).issubset(primary_index_cols):
@@ -111,7 +114,7 @@ class MultiGroupBy:
         This is the grouper created from the primary dataframe.
         """
         if not hasattr(self, "_primary_groupby"):
-            self._primary_groupby = self.primary_df.groupby(level=self.groupby, group_keys=False)
+            self._primary_groupby = self.primary_df.groupby(level=self.groupby, **self._kwargs)
         return self._primary_groupby
 
     @property
@@ -121,7 +124,7 @@ class MultiGroupBy:
         These are the groupers created from the secondary dataframes.
         """
         if not hasattr(self, "_secondary_groupbys"):
-            self._secondary_groupbys = [df.groupby(level=self.groupby, group_keys=False) for df in self.secondary_dfs]
+            self._secondary_groupbys = [df.groupby(level=self.groupby, **self._kwargs) for df in self.secondary_dfs]
         return self._secondary_groupbys
 
     def _get_secondary_vals(self, name: Union[str, tuple[str, ...]]) -> list[pd.DataFrame]:
@@ -180,6 +183,7 @@ def create_multi_groupby(
     primary_df: pd.DataFrame,
     secondary_dfs: Union[pd.DataFrame, list[pd.DataFrame]],
     groupby: Union[Hashable, list[str]],
+    **kwargs: Unpack[dict[str, Any]],
 ) -> MultiGroupBy:
     """Group multiple dataframes by the same index levels to apply a function to each group across all dataframes.
 
@@ -207,6 +211,8 @@ def create_multi_groupby(
         The secondary dataframes to group by.
     groupby
         The names of the index levels to group by.
+    kwargs
+        All further arguments will be passed to ``.groupby`` of all dataframes.
 
     Examples
     --------
@@ -255,7 +261,7 @@ def create_multi_groupby(
     Index: []
 
     """
-    return MultiGroupBy(primary_df, secondary_dfs, groupby)
+    return MultiGroupBy(primary_df, secondary_dfs, groupby, **kwargs)
 
 
 __all__ = ["sliding_window_view", "create_multi_groupby", "MultiGroupBy"]